// src/utils/enum.ts
export enum colors {
  primary = "is-primary",
  link = "is-link",
  info = "is-info",
  success = "is-success",
  warning = "is-warning",
  danger = "is-danger",
}

export enum buttonColors {
  primary = "is-primary",
  link = "is-link",
  info = "is-info",
  success = "is-success",
  warning = "is-warning",
  danger = "is-danger",
  white = "is-white",
  dark = "is-dark",
  black = "is-black",
  text = "is-text",
  ghost = "is-ghost",
}

export enum locations {
  topRight = "top-right",
  topCenter = "top-center",
  topLeft = "top-left",
  bottomRight = "bottom-right",
  bottomCenter = "bottom-center",
  bottomLeft = "bottom-left",
  middleRight = "middle-right",
  middleCenter = "middle-center",
  middleLeft = "middle-left",
}

export enum animations {
  fadeInTopLeft = "fadeInTopLeft",
  fadeInTopRight = "fadeInTopRight",
  fadeInBottomLeft = "fadeInBottomLeft",
  fadeInBottomRight = "fadeInBottomRight",
  backInDown = "backInDown",
  backInLeft = "backInLeft",
  backInRight = "backInRight",
  backInUp = "backInUp",
  bounceInDown = "bounceInDown",
  bounceInLeft = "bounceInLeft",
  bounceInRight = "bounceInRight",
  bounceInUp = "bounceInUp",
  fadeInDown = "fadeInDown",
  fadeInLeft = "fadeInLeft",
  fadeInRight = "fadeInRight",
  fadeInUp = "fadeInUp",
}

export enum sizes {
  small = "is-small",
  normal = "is-normal",
  medium = "is-medium",
  large = "is-large",
}

export enum align {
  right = "is-right",
  center = "is-center",
}

export enum inputStates {
  hovered = "is-hovered",
  focused = "is-focused",
  static = "is-static",
}

export enum buttonStates {
  hovered = "is-hovered",
  focused = "is-focused",
  static = "is-static",
  loading = "is-loading",
  active = "is-active",
}